subdomain to get SSL
Thank you for your excellent work.

The old hook.sh only works for domain.com, but does not work for sub.domain.com. This update adds the support to get SSL for subdomains. 

Sorry I'm a newbie for bash script. It would be great that someone can improve these codes.